+++ packages/dumb/src/utils/device/device.utils.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024-2025, Diabeloop
+ * Copyright (c) 2024-2026, Diabeloop
  *
  * All rights reserved.
  *