+++ packages/yourloops/test/integration/data/data-api.data.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023-2024, Diabeloop
+ * Copyright (c) 2023-2026, Diabeloop
  *
  * All rights reserved.
  *
@@ -26,255 +26,400 @@
  */
 
 import moment from 'moment-timezone'
-import { ChangeType, ParametersChange, Unit } from 'medical-domain'
+import { ChangeType, DblParameter, DeviceChangeName, DeviceHistory, ParametersChange, Unit } from 'medical-domain'
 
 export const history: ParametersChange[] = [
   {
     changeDate: '2022-11-01T00:00:00Z',
     parameters: [
       {
-        name: 'BOLUS_AGGRESSIVENESS_FACTOR',
+        name: DblParameter.AggressivenessHyperglycemia,
         value: '143',
         unit: Unit.Percent,
         level: 1,
         effectiveDate: '2022-11-01T00:00:00Z',
-        changeType: ChangeType.Added,
+        changeType: ChangeType.Added
       },
       {
-        name: 'LARGE_MEAL_BREAKFAST',
+        name: DblParameter.LargeBreakfast,
         value: '150',
         unit: Unit.Gram,
         level: 1,
         effectiveDate: '2022-11-01T00:00:00Z',
-        changeType: ChangeType.Added,
+        changeType: ChangeType.Added
       },
       {
-        name: 'LARGE_MEAL_DINNER',
+        name: DblParameter.LargeDinner,
         value: '150',
         unit: Unit.Gram,
         level: 1,
         effectiveDate: '2022-11-01T00:00:00Z',
-        changeType: ChangeType.Added,
+        changeType: ChangeType.Added
       },
       {
-        name: 'LARGE_MEAL_LUNCH',
+        name: DblParameter.LargeLunch,
         value: '70',
         unit: Unit.Gram,
         level: 1,
         effectiveDate: '2022-11-01T00:00:00Z',
-        changeType: ChangeType.Added,
+        changeType: ChangeType.Added
       },
       {
-        name: 'MEAL_RATIO_BREAKFAST_FACTOR',
+        name: DblParameter.AggressivenessBreakfast,
         value: '100',
         unit: Unit.Percent,
         level: 1,
         effectiveDate: '2022-11-01T00:00:00Z',
-        changeType: ChangeType.Added,
+        changeType: ChangeType.Added
       },
       {
-        name: 'MEAL_RATIO_BREAKFAST_FACTOR',
+        name: DblParameter.AggressivenessBreakfast,
         value: '110',
         unit: Unit.Percent,
         level: 1,
         effectiveDate: '2022-11-02T17:00:40Z',
         changeType: ChangeType.Updated,
         previousValue: '100',
-        previousUnit: Unit.Percent,
+        previousUnit: Unit.Percent
       },
       {
-        name: 'MEAL_RATIO_BREAKFAST_FACTOR',
+        name: DblParameter.AggressivenessBreakfast,
         value: '100',
         unit: Unit.Percent,
         level: 1,
         effectiveDate: '2022-11-01T00:00:00Z',
         changeType: ChangeType.Updated,
         previousValue: '110',
-        previousUnit: Unit.Percent,
+        previousUnit: Unit.Percent
       },
       {
-        name: 'MEAL_RATIO_DINNER_FACTOR',
+        name: DblParameter.AggressivenessDinner,
         value: '100',
         unit: Unit.Percent,
         level: 1,
         effectiveDate: '2022-11-01T00:00:00Z',
-        changeType: ChangeType.Added,
+        changeType: ChangeType.Added
       },
       {
-        name: 'MEAL_RATIO_DINNER_FACTOR',
+        name: DblParameter.AggressivenessDinner,
         value: '90',
         unit: Unit.Percent,
         level: 1,
         effectiveDate: '2022-11-02T17:00:40Z',
         changeType: ChangeType.Updated,
         previousValue: '100',
-        previousUnit: Unit.Percent,
+        previousUnit: Unit.Percent
       },
       {
-        name: 'MEAL_RATIO_DINNER_FACTOR',
+        name: DblParameter.AggressivenessDinner,
         value: '100',
         unit: Unit.Percent,
         level: 1,
         effectiveDate: '2022-11-01T00:00:00Z',
         changeType: ChangeType.Updated,
         previousValue: '90',
-        previousUnit: Unit.Percent,
+        previousUnit: Unit.Percent
       },
       {
-        name: 'MEAL_RATIO_LUNCH_FACTOR',
+        name: DblParameter.AggressivenessLunch,
         value: '130',
         unit: Unit.Percent,
         level: 1,
         effectiveDate: '2022-11-01T00:00:00Z',
-        changeType: ChangeType.Added,
+        changeType: ChangeType.Added
       },
       {
-        name: 'MEAL_RATIO_LUNCH_FACTOR',
+        name: DblParameter.AggressivenessLunch,
         value: '90',
         unit: Unit.Percent,
         level: 1,
         effectiveDate: '2022-11-07T14:01:14Z',
         changeType: ChangeType.Updated,
         previousValue: '130',
-        previousUnit: Unit.Percent,
+        previousUnit: Unit.Percent
       },
       {
-        name: 'MEDIUM_MEAL_BREAKFAST',
+        name: DblParameter.AverageBreakfast,
         value: '70',
         unit: Unit.Gram,
         level: 1,
         effectiveDate: '2022-11-01T00:00:00Z',
-        changeType: ChangeType.Added,
+        changeType: ChangeType.Added
       },
       {
-        name: 'MEDIUM_MEAL_DINNER',
+        name: DblParameter.AverageDinner,
         value: '60',
         unit: Unit.Gram,
         level: 1,
         effectiveDate: '2022-11-01T00:00:00Z',
-        changeType: ChangeType.Added,
+        changeType: ChangeType.Added
       },
       {
-        name: 'MEDIUM_MEAL_LUNCH',
+        name: DblParameter.AverageLunch,
         value: '50',
         unit: Unit.Gram,
         level: 1,
         effectiveDate: '2022-11-01T00:00:00Z',
-        changeType: ChangeType.Added,
+        changeType: ChangeType.Added
       },
       {
-        name: 'PATIENT_BASAL_AGGRESSIVENESS_FACTOR_LEVEL_IN_EUGLYCAEMIA',
+        name: DblParameter.AggressivenessNormoglycemia,
         value: '100',
         unit: Unit.Percent,
         level: 1,
         effectiveDate: '2022-11-01T00:00:00Z',
-        changeType: ChangeType.Added,
+        changeType: ChangeType.Added
       },
       {
-        name: 'PATIENT_GLYCEMIA_TARGET',
+        name: DblParameter.TargetGlucoseLevel,
         value: '100.0',
         unit: Unit.MilligramPerDeciliter,
         level: 1,
         effectiveDate: '2022-11-01T00:00:00Z',
-        changeType: ChangeType.Added,
+        changeType: ChangeType.Added
       },
       {
-        name: 'PATIENT_GLY_HYPER_LIMIT',
+        name: DblParameter.HyperglycemiaThreshold,
         value: '180.1',
         unit: Unit.MilligramPerDeciliter,
         level: 1,
         effectiveDate: '2022-11-01T00:00:00Z',
-        changeType: ChangeType.Added,
+        changeType: ChangeType.Added
       },
       {
-        name: 'PATIENT_GLY_HYPER_LIMIT',
+        name: DblParameter.HyperglycemiaThreshold,
         value: '140',
         unit: Unit.MilligramPerDeciliter,
         level: 1,
         effectiveDate: '2022-11-02T07:00:00Z',
         changeType: ChangeType.Updated,
         previousValue: '180.1',
-        previousUnit: Unit.MilligramPerDeciliter,
+        previousUnit: Unit.MilligramPerDeciliter
       },
       {
-        name: 'PATIENT_GLY_HYPER_LIMIT',
+        name: DblParameter.HyperglycemiaThreshold,
         value: '180.1',
         unit: Unit.MilligramPerDeciliter,
         level: 1,
         effectiveDate: '2022-11-01T00:00:00Z',
         changeType: ChangeType.Updated,
         previousValue: '140',
-        previousUnit: Unit.MilligramPerDeciliter,
+        previousUnit: Unit.MilligramPerDeciliter
       },
       {
-        name: 'PATIENT_GLY_HYPO_LIMIT',
+        name: DblParameter.HypoglycemiaThreshold,
         value: '70',
         unit: Unit.MilligramPerDeciliter,
         level: 1,
         effectiveDate: '2022-11-01T00:00:00Z',
-        changeType: ChangeType.Added,
+        changeType: ChangeType.Added
       },
       {
-        name: 'PATIENT_GLY_HYPO_LIMIT',
+        name: DblParameter.HypoglycemiaThreshold,
         value: '60',
         unit: Unit.MilligramPerDeciliter,
         level: 1,
         effectiveDate: '2022-11-02T07:00:00Z',
         changeType: ChangeType.Updated,
         previousValue: '70',
-        previousUnit: Unit.MilligramPerDeciliter,
+        previousUnit: Unit.MilligramPerDeciliter
       },
       {
-        name: 'PATIENT_GLY_HYPO_LIMIT',
+        name: DblParameter.HypoglycemiaThreshold,
         value: '70',
         unit: Unit.MilligramPerDeciliter,
         level: 1,
         effectiveDate: '2022-11-01T00:00:00Z',
         changeType: ChangeType.Updated,
         previousValue: '60',
-        previousUnit: Unit.MilligramPerDeciliter,
+        previousUnit: Unit.MilligramPerDeciliter
       },
       {
-        name: 'SMALL_MEAL_BREAKFAST',
+        name: DblParameter.SmallBreakfast,
         value: '15',
         unit: Unit.Gram,
         level: 1,
         effectiveDate: '2022-11-01T00:00:00Z',
-        changeType: ChangeType.Added,
+        changeType: ChangeType.Added
       },
       {
-        name: 'SMALL_MEAL_DINNER',
+        name: DblParameter.SmallDinner,
         value: '20',
         unit: Unit.Gram,
         level: 1,
         effectiveDate: '2022-11-01T00:00:00Z',
-        changeType: ChangeType.Added,
+        changeType: ChangeType.Added
       },
       {
-        name: 'SMALL_MEAL_LUNCH',
+        name: DblParameter.SmallLunch,
         value: '30',
         unit: Unit.Gram,
         level: 1,
         effectiveDate: '2022-11-01T00:00:00Z',
-        changeType: ChangeType.Added,
+        changeType: ChangeType.Added
       },
       {
-        name: 'TOTAL_INSULIN_FOR_24H',
+        name: DblParameter.TotalDailyInsulin,
         value: '53',
         unit: Unit.InsulinUnit,
         level: 1,
         effectiveDate: '2022-11-01T00:00:00Z',
-        changeType: ChangeType.Added,
+        changeType: ChangeType.Added
       },
       {
-        name: 'WEIGHT',
+        name: DblParameter.Weight,
         value: '69.0',
         unit: Unit.Kilogram,
         level: 1,
         effectiveDate: '2022-11-01T00:00:00Z',
-        changeType: ChangeType.Added,
+        changeType: ChangeType.Added
       }
     ]
+  }
+]
+
+export const devicesHistory: DeviceHistory[] = [
+  {
+    changeDate: "2019-04-12T00:02Z",
+    devices:
+      [
+        {
+          changeType: ChangeType.Added,
+          effectiveDate: "2019-04-12T00:02:00.000Z",
+          name: DeviceChangeName.CgmManufacturer,
+          value: "Dexcom"
+        },
+        {
+          changeType: ChangeType.Added,
+          effectiveDate: "2019-04-12T00:02:00.000Z",
+          name: DeviceChangeName.CgmName,
+          value: "G5"
+        },
+        {
+          changeType: ChangeType.Added,
+          effectiveDate: "2019-04-12T00:02:00.000Z",
+          name: DeviceChangeName.HandsetDeviceId,
+          value: "MobiGo+"
+        },
+        {
+          changeType: ChangeType.Added,
+          effectiveDate: "2019-04-12T00:02:00.000Z",
+          name: DeviceChangeName.HandsetImei,
+          value: "0123456789"
+        },
+        {
+          changeType: ChangeType.Added,
+          effectiveDate: "2019-04-12T00:02:00.000Z",
+          name: DeviceChangeName.HandsetSoftwareVersion,
+          value: "1.12.12-DBLG1-INS-DEXG6-COMMERCIAL"
+        },
+        {
+          changeType: ChangeType.Added,
+          effectiveDate: "2019-04-12T00:02:00.000Z",
+          name: DeviceChangeName.MobileAppActivationCode,
+          value: "123-456-789"
+        },
+        {
+          changeType: ChangeType.Added,
+          effectiveDate: "2019-04-12T00:02:00.000Z",
+          name: DeviceChangeName.MobileAppVersion,
+          value: "1.0.0"
+        },
+        {
+          changeType: ChangeType.Added,
+          effectiveDate: "2019-04-12T00:02:00.000Z",
+          name: DeviceChangeName.PumpManufacturer,
+          value: "ROCHE"
+        },
+        {
+          changeType: ChangeType.Added,
+          effectiveDate: "2019-04-12T00:02:00.000Z",
+          name: DeviceChangeName.PumpName,
+          value: "Insight"
+        },
+        {
+          changeType: ChangeType.Added,
+          effectiveDate: "2019-04-12T00:02:00.000Z",
+          name: DeviceChangeName.SmartphoneModel,
+          value: "A212F"
+        }
+      ]
+  },
+  {
+    changeDate: "2019-04-26T00:02Z",
+    devices:
+      [
+        {
+          changeType: ChangeType.Added,
+          effectiveDate: "2019-04-12T00:02:00.000Z",
+          name: DeviceChangeName.CgmManufacturer,
+          value: "Abbott",
+          previousValue: "Dexcom"
+        },
+        {
+          changeType: ChangeType.Updated,
+          effectiveDate: "2019-04-26T00:02:00.000Z",
+          name: DeviceChangeName.CgmName,
+          value: "G6",
+          previousValue: "G5"
+        },
+        {
+          changeType: ChangeType.Added,
+          effectiveDate: "2019-04-12T00:02:00.000Z",
+          name: DeviceChangeName.HandsetDeviceId,
+          value: "DH22",
+          previousValue: "MobiGo+"
+        },
+        {
+          changeType: ChangeType.Added,
+          effectiveDate: "2019-04-12T00:02:00.000Z",
+          name: DeviceChangeName.HandsetImei,
+          value: "111111111",
+          previousValue: "0123456789"
+        },
+        {
+          changeType: ChangeType.Added,
+          effectiveDate: "2019-04-12T00:02:00.000Z",
+          name: DeviceChangeName.HandsetSoftwareVersion,
+          value: "1.18.0.1300-DBLG1-KAL-G6-COMMERCIAL",
+          previousValue: "1.12.12-DBLG1-INS-DEXG6-COMMERCIAL"
+        },
+        {
+          changeType: ChangeType.Added,
+          effectiveDate: "2019-04-12T00:02:00.000Z",
+          name: DeviceChangeName.MobileAppActivationCode,
+          value: "111-111-111",
+          previousValue: "123-456-789"
+        },
+        {
+          changeType: ChangeType.Updated,
+          effectiveDate: "2019-04-26T00:02:00.000Z",
+          name: DeviceChangeName.MobileAppVersion,
+          value: "1.1.0",
+          previousValue: "1.0.0"
+        },
+        {
+          changeType: ChangeType.Updated,
+          effectiveDate: "2019-04-26T00:02:00.000Z",
+          name: DeviceChangeName.PumpManufacturer,
+          value: "VICENTRA",
+          previousValue: "ROCHE"
+        },
+        {
+          changeType: ChangeType.Updated,
+          effectiveDate: "2019-04-26T00:02:00.000Z",
+          name: DeviceChangeName.PumpName,
+          value: "Kaleido",
+          previousValue: "Insight"
+        },
+        {
+          changeType: ChangeType.Updated,
+          effectiveDate: "2019-04-26T00:02:00.000Z",
+          name: DeviceChangeName.SmartphoneModel,
+          value: "A26",
+          previousValue: "A212F"
+        }
+      ]
   }
 ]