+++ packages/yourloops/test/integration/use-cases/device-settings-visualisation.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023-2025, Diabeloop
+ * Copyright (c) 2023-2026, Diabeloop
  *
  * All rights reserved.
  *
@@ -28,18 +28,33 @@
 import {
   checkChangeHistoryContent,
   checkCopyTextButton,
-  checkCurrentParametersContent,
+  checkCurrentParametersContent, checkDeviceHistoryContent,
   checkDevicesMenuLayoutWithBasalSafetyProfile,
   checkDevicesMenuLayoutWithoutBasalSafetyProfile,
   checkG2CopyTextButton,
-  checkG2CurrentParametersContent,
+  checkG2CurrentParametersContent, checkParametersMemo, checkParametersMemoNotDisplayed,
   checkSafetyBasalProfileContent,
   checkSafetyBasalProfileErrorMessage
 } from '../assert/device-view.assert'
 
+export const testDevicesVisualisationForHcp = async () => {
+  testDevicesMenuLayout()
+  await testCurrentParametersVisualisationForHcp()
+  await testBasalSafetyProfileVisualisation()
+  await testChangeHistoryVisualisation()
+}
+
 export const testDevicesVisualisation = async () => {
   testDevicesMenuLayout()
   await testCurrentParametersVisualisation()
+  await testBasalSafetyProfileVisualisation()
+  await testChangeHistoryVisualisation()
+  await testDeviceHistoryVisualisation()
+}
+
+export const testG2DevicesVisualisationForHcp = async () => {
+  testDevicesMenuLayout()
+  await testG2CurrentParametersVisualisationForHcp()
   await testBasalSafetyProfileVisualisation()
   await testChangeHistoryVisualisation()
 }
@@ -49,6 +64,7 @@ export const testG2DevicesVisualisation = async () => {
   await testG2CurrentParametersVisualisation()
   await testBasalSafetyProfileVisualisation()
   await testChangeHistoryVisualisation()
+  await testDeviceHistoryVisualisation()
 }
 
 export const testEmptySafetyBasalProfileGenericErrorMessage = async () => {
@@ -63,13 +79,27 @@ export const testDevicesMenuLayoutForMobigoDevice = () => {
   checkDevicesMenuLayoutWithoutBasalSafetyProfile()
 }
 
+const testCurrentParametersVisualisationForHcp = async () => {
+  checkCurrentParametersContent()
+  await checkParametersMemo()
+  await checkCopyTextButton()
+}
+
 const testCurrentParametersVisualisation = async () => {
   checkCurrentParametersContent()
+  checkParametersMemoNotDisplayed()
   await checkCopyTextButton()
+}
+
+const testG2CurrentParametersVisualisationForHcp = async () => {
+  checkG2CurrentParametersContent()
+  await checkParametersMemo()
+  await checkG2CopyTextButton()
 }
 
 const testG2CurrentParametersVisualisation = async () => {
   checkG2CurrentParametersContent()
+  checkParametersMemoNotDisplayed()
   await checkG2CopyTextButton()
 }
 
@@ -79,4 +109,8 @@ const testBasalSafetyProfileVisualisation = async () => {
 
 const testChangeHistoryVisualisation = async () => {
   await checkChangeHistoryContent()
+}
+
+const testDeviceHistoryVisualisation = async () => {
+  await checkDeviceHistoryContent()
 }